Making NOT IN behave like IN special behaviour
